# api/openapi.yaml
openapi: 3.0.3
info:
  title: twofa API
  version: 0.2.0
  description: |
    Сервер двухфакторной аутентификации twofa: коды доставки (email / SMS /
    Telegram / Telegram-push), TOTP-приложения, резервные коды и WebAuthn
    (passkeys). Четыре группы REST-эндпоинтов: публичный API входа (без
    состояния, для интеграций), web-сессии (cookie + CSRF), кабинет
    пользователя и админ API (Bearer admin_token).

    Поддерживаемые факторы: TOTP (RFC 6238), коды доставки по каналам
    email/sms/telegram (и approval push в Telegram), одноразовые резервные
    коды, WebAuthn/passkeys.

    Аутентификация: публичные /api/v1/auth/* — без авторизации; /api/v1/me/*
    и /api/v1/logout — cookie twofa_session (+ заголовок X-CSRF-Token на
    мутациях); /api/v1/admin/* — Authorization: Bearer <admin_token>.

    Ошибки — JSON вида {"error":"<машинный_код>"}. Коды ответов значимы:
    401 bad_credentials / bad_code, 409 no_channel / no_credentials /
    username_taken, 423 locked (fail-блокировка), 429 rate_limited /
    cooldown (+ retry_after, заголовок Retry-After), 410 expired.

    RADIUS-интерфейс (Access-Request/Accounting на :1812/:1813, для
    MikroTik/RouterOS) в REST не документируется — см. README.

    Спецификация и человекочитаемая страница: GET /openapi.yaml и /api/docs.
servers:
  - url: http://localhost:8080
    description: Локальный сервер (пример)
...
security: {}
paths:
...
  /oidc/authorize:
    get:
      tags: [OIDC]
      summary: "Authorization endpoint (вход через Ligament)"
      description: "Проверяет client_id и redirect_uri (строгое совпадение; ошибки клиента — страница 400 без редиректа), затем параметры запроса (ошибки — редирект с error/state). Без web-сессии — редирект /login?next=<authorize>; с сессией — страница согласия (oidc_consent), подтверждение выпускает код. Параметры: client_id, redirect_uri, response_type=code, scope (обязателен openid), state, nonce, code_challenge+code_challenge_method (S256 обязателен для public-клиентов)."
      operationId: oidcAuthorize
      parameters:
        - name: client_id
          in: query
          required: true
          schema: {type: string}
        - name: redirect_uri
          in: query
          required: true
          schema: {type: string}
        - name: response_type
          in: query
          required: true
          schema: {type: string, enum: [code]}
        - name: scope
          in: query
          required: true
          schema: {type: string, example: "openid profile email"}
        - name: state
          in: query
          schema: {type: string}
        - name: nonce
          in: query
          schema: {type: string}
        - name: code_challenge
          in: query
          schema: {type: string}
        - name: code_challenge_method
          in: query
          schema: {type: string, enum: [S256, plain]}
      responses:
        '200':
          description: Страница согласия (HTML).
        '302':
          description: "redirect_uri?code=…&state=… (после подтверждения), /login?next=… (без сессии) либо redirect_uri?error=… (invalid_scope, unsupported_response_type, invalid_request)."
        '400':
          description: "Страница ошибки (HTML): неизвестный client_id или незарегистрированный redirect_uri — редиректить нельзя."
  /oidc/authorize/confirm:
    post:
      tags: [OIDC]
      summary: "Подтверждение согласия (HTML-форма)"
      description: "POST со страницы согласия: CSRF-токен сессии + скрытые поля всех параметров authorize. Выпускает authorization-код (32 байта, TTL 60 с, одноразовый) и редиректит на redirect_uri?code=…&state=…. Аудит oidc_consent."
      operationId: oidcAuthorizeConfirm
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [csrf_token, client_id, redirect_uri, response_type, scope]
              properties:
                csrf_token: {type: string}
                client_id: {type: string}
                redirect_uri: {type: string}
                response_type: {type: string}
                scope: {type: string}
                state: {type: string}
                nonce: {type: string}
                code_challenge: {type: string}
                code_challenge_method: {type: string}
      responses:
        '302':
          description: "redirect_uri?code=…&state=…; без сессии — /login?next=…; ошибка запроса — redirect_uri?error=…."
        '400':
          description: "Страница ошибки (HTML): неизвестный клиент/redirect_uri."
        '403':
          description: "Отсутствует или неверен CSRF-токен (text/plain)."
  /oidc/token:
    post:
      tags: [OIDC]
      summary: "Token endpoint (grant_type=authorization_code)"
      description: "Аутентификация клиента: Basic или form (public — без секрета, PKCE обязателен). Атомарно погашает код (повторное использование/истёкший → invalid_grant), проверяет PKCE (S256/plain) и выдаёт access-токен (TTL 300 с) и подписанный ID-токен (RS256, TTL 5 мин). Refresh-токенов нет. Аудит oidc_token. Ошибки — RFC 6749 §5.2: {error}."
      operationId: oidcToken
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [grant_type, code, redirect_uri]
              properties:
                grant_type: {type: string, enum: [authorization_code]}
                code: {type: string}
                redirect_uri: {type: string}
                client_id: {type: string}
                client_secret: {type: string, description: "form-аутентификация (альтернатива Basic)."}
                code_verifier: {type: string, description: "PKCE; обязателен, если код выдан с code_challenge."}
            example: {grant_type: authorization_code, code: "…", redirect_uri: "https://grafana.example.com/login/generic_oauth", code_verifier: "dBjftJeZ4CVP…"}
      responses:
        '200':
          description: "Токены; ID-токен содержит iss/sub/aud/exp/iat/auth_time/amr/groups (+nonce, profile/email по scope)."
          content:
            application/json:
              schema:
                type: object
                required: [access_token, token_type, expires_in, scope, id_token]
                properties:
                  access_token: {type: string}
                  token_type: {type: string, enum: [Bearer]}
                  expires_in: {type: integer, example: 300}
                  scope: {type: string}
                  id_token: {type: string, description: "JWT RS256."}
              example: {access_token: "…", token_type: Bearer, expires_in: 300, scope: "openid profile email", id_token: "eyJhbGciOiJSUzI1NiIs…"}
        '400':
          description: "invalid_request / invalid_grant / unsupported_grant_type."
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
              example: {error: invalid_grant}
        '401':
          description: invalid_client (неверный/неизвестный клиент).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
              example: {error: invalid_client}
  /oidc/userinfo:
    get:
      tags: [OIDC]
      summary: "Клеймы пользователя по access-токену"
      description: "Authorization: Bearer <access_token> (из /oidc/token, TTL 300 с). sub обязателен; preferred_username/name/groups — по scope profile, email/email_verified — по scope email."
      operationId: oidcUserinfo
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Клеймы пользователя.
          content:
            application/json:
              schema:
                type: object
                required: [sub]
                properties:
                  sub: {type: string, description: "UUID пользователя."}
                  preferred_username: {type: string}
                  name: {type: string}
                  groups: {type: array, items: {type: string}}
                  email: {type: string}
                  email_verified: {type: boolean}
              example: {sub: "11111111-1111-1111-1111-111111111111", preferred_username: vasya, name: Василий, groups: [user], email: vasya@example.com, email_verified: true}
        '401':
          description: "invalid_token: токен отсутствует, неизвестен или истёк."
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
              example: {error: invalid_token}
components:
  securitySchemes:
    adminBearer:
      type: http
      scheme: bearer
      description: "Админ-токен из настроек (admin_token; задать/перегенерировать — POST /api/v1/admin/settings/regenerate). Заголовок Authorization: Bearer <token>; сравнение SHA-256 в постоянном времени."
    bearerAuth:
      type: http
      scheme: bearer
      description: "Access-токен /oidc/token (TTL 300 с); заголовок Authorization: Bearer <access_token> для /oidc/userinfo."
    sessionCookie:
      type: apiKey
      in: cookie
      name: twofa_session
      description: "Web-сессия (cookie twofa_session, HttpOnly+Secure+SameSite=Lax; выдаётся POST /api/v1/login и /api/v1/login/2fa). Для МУТИРУЮЩИХ методов (POST/PUT/PATCH/DELETE) обязателен заголовок X-CSRF-Token со значением csrf из ответа входа — иначе 403 csrf. Доверенные устройства — отдельная cookie twofa_device (remember_device)."
  schemas:
...
    Error:
      type: object
      required: [error]
      description: "Машинный код ошибки без внутренних деталей."
      properties:
        error:
          type: string
      example: {error: bad_credentials}
...